requirements.txt
requests>=2.31.0
dnspython>=2.4.0
pyyaml>=6.0
tqdm>=4.66.0
